--- .gitignore
@@ -1,5 +1,0 @@
-node_modules
-dist
-npm-debug.log
-.env
-.DS_Store
