--- .travis.yml
@@ -1,6 +1,0 @@
-language: node_js
-node_js:
-  - "stable"
-install:
-  - npm install
-  - npm run build
