+++ README.md
@@ -1,19 +1,7 @@
 # Gubongo
 Csodálatos FéregTv - chatjét vizualizáló web alkalmazás. Egy websocket szolgáltatáshoz csatlakozva megjeleníti a csatorna chatjére író felhasználók "kukacát".
 
-# Install
-```sh
+# install
 npm install
-```
 
-# Build
-```sh
-npm run build
-```
-Open dist/index.html
-
-# Development
-```sh
-npm run start
-```
-Open http://localhost:8080
+utána csak menyitod az index.html-t
