+++ index.html
@@ -31,5 +31,7 @@
   </head>
   <body>
     <img src="images/background/bg.jpg" id="background-image" class="background"/>
+    <script type="text/javascript" src="engine.js"></script>
+    <script type="text/javascript" src="gubongo_v.0.5.js?v=190105"></script>
   </body>
 </html>
