+++ package.json
@@ -4,17 +4,8 @@
   "description": "",
   "main": "gubongo.js",
   "dependencies": {},
-  "devDependencies": {
-    "clean-webpack-plugin": "^1.0.0",
-    "copy-webpack-plugin": "^4.6.0",
-    "html-webpack-plugin": "^3.2.0",
-    "webpack": "4.28.3",
-    "webpack-cli": "3.2.0",
-    "webpack-dev-server": "3.1.14"
-  },
+  "devDependencies": {},
   "scripts": {
-    "build": "webpack --config webpack.prod.js",
-    "start": "webpack-dev-server --config webpack.dev.js --hot",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
