--- webpack.prod.js
@@ -1,20 +1,0 @@
-const path = require('path');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-
-module.exports = {
-    mode: 'development',
-    entry: './src/gubongo',
-    plugins: [
-        new CleanWebpackPlugin(['dist']),
-        new HtmlWebpackPlugin({
-         template: 'index.html',
-        }),
-        new CopyWebpackPlugin([ { from: 'images', to: 'images' } ]),
-      ],
-    output: {
-        path: path.resolve(__dirname, 'dist'),
-        filename: '[name].[contenthash].js',
-    },
-};
